Done with the project
Implemented the javascript that changes the color of the boxes when
the mouse enter the box. Made sure the size of the box remained the same
even though the amount of box increase. Also implemented the prompt method
that asks the user for the size and then the size changes based on the
user's input.